feat: add components loading and DetailsInfo
