+++ src/components/DetailsInfo/types.ts
@@ -1,0 +1,9 @@
+import React from 'react';
+
+interface IDetailsInfoProps {
+  id: string;
+  handleClickClose: () => void;
+  children: React.ReactNode;
+}
+
+export type { IDetailsInfoProps };
