+++ src/components/Loading/loading.module.scss
@@ -1,0 +1,12 @@
+@import '../../styles/settings/index';
+
+.loader {
+  @include sizes(120px, 120px);
+  margin: 0 auto;
+}
+
+@media (max-width: $mob) {
+  .loader {
+    @include sizes(60px, 60px);
+  }
+}
